login/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
motor # MongoDB 비동기 드라이버
sendgrid